[Fix] 28 컨벤션 수정
